Added many reset data functions to mapeditor  cargomovecostdivisor for vehicles is now float  Fixed: objects could not be attacked  Filenames in cache

+++ source/vehicletype.h
@@ -147,7 +147,7 @@ extern const char*  cvehiclefunctions[];
         int          height;
 
         //! if a transport moves the movement for the units inside a transport is decreased by 1/n of the tranport's distance
-        int     cargoMovementDivisor;
+        double     cargoMovementDivisor;
 
         // if the unit can change the level of height, this is the number of fields the unit must move to go from one level to the next
         // int          steigung;
